// Frontend-Web/src/app/model/usuario/usuario.model.ts
export interface Usuario {
    idUsuario?: string;
    cuitUsuario: number;
    nombreUsuario: string;
    apellidoUsuario: string;
    contrasenaUsuario: string;
    dniUsuario: number;
    domicilioUsuario: string;
    emailUsuario: string;
    idDepartamento: number;
    nroCelularUsuario: number;
    nroTelefonoUsuario: number;
    idRol?: number;
    idEstadoUsuario?: number
  }